move getCollator(type) from WriteCollationData to UCA
