+++ unicodetools/src/main/java/org/unicode/text/UCA/UCA.java
@@ -107,8 +107,33 @@ public final class UCA implements Comparator<String>, UCA_Types {
         cldr,
     }
 
-    public static final String copyright =
-            "Copyright (C) 2000, IBM Corp. and others. All Rights Reserved.";
+    private static UCA ducetCollator;
+    private static UCA cldrCollator;
+
+    public static UCA getDucetCollator() {
+        if (ducetCollator == null) {
+            ducetCollator = buildDucetCollator();
+        }
+        return ducetCollator;
+    }
+
+    public static UCA getCldrCollator() {
+        if (cldrCollator == null) {
+            cldrCollator = buildCldrCollator();
+        }
+        return cldrCollator;
+    }
+
+    public static UCA getCollator(CollatorType type) {
+        switch (type) {
+            case cldr:
+                return getCldrCollator();
+            case ducet:
+                return getDucetCollator();
+            default:
+                throw new IllegalArgumentException();
+        }
+    }
 
     @Override
     public int compare(String a, String b) {
@@ -1797,7 +1822,7 @@ public final class UCA implements Comparator<String>, UCA_Types {
         return buildCollator(-1, -1);
     }
 
-    static UCA buildCollator(int variableHigh, int firstNonVariable) {
+    private static UCA buildCollator(int variableHigh, int firstNonVariable) {
         try {
             if (VERBOSE) System.out.println("Building UCA");
             final Path dataPath = Settings.UnicodeTools.getDataPathForLatestVersion("uca");
@@ -1815,6 +1840,116 @@ public final class UCA implements Comparator<String>, UCA_Types {
         } catch (final IOException e) {
             throw new IllegalArgumentException(e);
         }
+    }
+
+    private static UCA buildCldrCollator() {
+        final UCA ducet = getDucetCollator();
+
+        final int ducetVariableHigh = CEList.getPrimary(ducet.getVariableHighCE());
+        int cldrVariableHigh = 0;
+        // This will normally come out as ducetVariableHigh + 1.
+        int firstDucetNonVariable = -1;
+
+        // DUCET: variable primary weights include spaces, punctuation, and symbols
+        // CLDR: variable primary weights include only spaces and punctuation
+        for (final UCA.Primary up : ducet.getRegularPrimaries()) {
+            final int ducetPrimary = up.primary;
+            if (firstDucetNonVariable < 0 && ducetPrimary > ducetVariableHigh) {
+                firstDucetNonVariable = ducetPrimary;
+            }
+
+            final String repChar = up.getRepresentative();
+            final int firstChar = Character.codePointAt(repChar, 0);
+            final int cat = Default.ucd().getCategory(firstChar);
+            switch (cat) {
+                case UCD_Types.DASH_PUNCTUATION:
+                case UCD_Types.START_PUNCTUATION:
+                case UCD_Types.END_PUNCTUATION:
+                case UCD_Types.CONNECTOR_PUNCTUATION:
+                case UCD_Types.OTHER_PUNCTUATION:
+                case UCD_Types.INITIAL_PUNCTUATION:
+                case UCD_Types.FINAL_PUNCTUATION:
+                    if (ducetPrimary <= ducetVariableHigh && ducetPrimary > cldrVariableHigh) {
+                        cldrVariableHigh = ducetPrimary;
+                    }
+                    break;
+                default:
+                    break;
+            }
+        }
+        final UCA result = buildCollator(cldrVariableHigh, firstDucetNonVariable);
+
+        result.overrideCE("\uFFFE", 0x1, 0x20, 2);
+        result.overrideCE("\uFFFF", 0xFFFE, 0x20, 2);
+
+        if (
+        /* ADD_TIBETAN */ true) {
+            final CEList fb2 = result.getCEList("\u0FB2", true);
+            final CEList fb3 = result.getCEList("\u0FB3", true);
+            final CEList f71_f72 = result.getCEList("\u0F71\u0F72", true);
+            final CEList f71_f74 = result.getCEList("\u0F71\u0F74", true);
+            final CEList fb2_f71 = result.getCEList("\u0FB2\u0F71", true);
+            final CEList fb3_f71 = result.getCEList("\u0FB3\u0F71", true);
+
+            addOverride(
+                    result,
+                    "\u0FB2\u0F71",
+                    fb2_f71); // 0FB2 0F71      ;     [.255A.0020.0002.0FB2][.2570.0020.0002.0F71] -
+            // concat 0FB2 + 0F71
+            addOverride(
+                    result,
+                    "\u0FB2\u0F71\u0F72",
+                    fb2,
+                    f71_f72); // 0FB2 0F71 0F72 ;    [.255A.0020.0002.0FB2][.2572.0020.0002.0F73] -
+            // concat 0FB2 + (0F71/0F72)
+            addOverride(result, "\u0FB2\u0F73", fb2, f71_f72); // 0FB2 0F73      ;
+            // [.255A.0020.0002.0FB2][.2572.0020.0002.0F73] = prev
+            addOverride(
+                    result,
+                    "\u0FB2\u0F71\u0F74",
+                    fb2,
+                    f71_f74); // 0FB2 0F71 0F74 ;    [.255A.0020.0002.0FB2][.2576.0020.0002.0F75] -
+            // concat 0FB2 + (0F71/0F74)
+            addOverride(result, "\u0FB2\u0F75", fb2, f71_f74); // 0FB2 0F75      ;
+            // [.255A.0020.0002.0FB2][.2576.0020.0002.0F75]  = prev
+
+            // same as above, but 0FB2 => 0FB3 and fb2 => fb3
+
+            addOverride(
+                    result,
+                    "\u0FB3\u0F71",
+                    fb3_f71); // 0FB3 0F71      ;     [.255A.0020.0002.0FB3][.2570.0020.0002.0F71] -
+            // concat 0FB3 + 0F71
+            addOverride(
+                    result,
+                    "\u0FB3\u0F71\u0F72",
+                    fb3,
+                    f71_f72); // 0FB3 0F71 0F72 ;    [.255A.0020.0002.0FB3][.2572.0020.0002.0F73] -
+            // concat 0FB3 + (0F71/0F72)
+            addOverride(result, "\u0FB3\u0F73", fb3, f71_f72); // 0FB3 0F73      ;
+            // [.255A.0020.0002.0FB3][.2572.0020.0002.0F73] = prev
+            addOverride(
+                    result,
+                    "\u0FB3\u0F71\u0F74",
+                    fb3,
+                    f71_f74); // 0FB3 0F71 0F74 ;    [.255A.0020.0002.0FB3][.2576.0020.0002.0F75] -
+            // concat 0FB3 + (0F71/0F74)
+            addOverride(result, "\u0FB3\u0F75", fb3, f71_f74); // 0FB3 0F75      ;
+            // [.255A.0020.0002.0FB3][.2576.0020.0002.0F75]  = prev
+        }
+
+        return result;
+    }
+
+    private static void addOverride(UCA result, String string, CEList... ceLists) {
+        final IntStack tempStack = new IntStack(10);
+        for (final CEList ceList : ceLists) {
+            for (int i = 0; i < ceList.length(); ++i) {
+                final int ce = ceList.at(i);
+                tempStack.append(ce);
+            }
+        }
+        result.overrideCE(string, tempStack);
     }
 
     UCA_Statistics getStatistics() {
